pyproject.toml
[tool.poetry]
name = "ecommerce-analytics"
version = "0.1.0"
description = ""
authors = ["Luciano Filho <lvgalvaofilho@gmail.com>"]
readme = "README.md"
packages = [{ include = "ecommerce_analytics" }]

[tool.poetry.dependencies]
python = "^3.11"
pyspark = "^3.5.0"
delta-spark = "^3.0.0"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.12.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
